Moving utilities to a subpackage
