+++ src/main/java/edu/clemson/cs/rsrg/vcgeneration/proofrules/declaration/GenericTypeVariableDeclRule.java
@@ -13,10 +13,10 @@
 package edu.clemson.cs.rsrg.vcgeneration.proofrules.declaration;
 
 import edu.clemson.cs.rsrg.absyn.declarations.variabledecl.VarDec;
-import edu.clemson.cs.rsrg.vcgeneration.Utilities;
 import edu.clemson.cs.rsrg.vcgeneration.absyn.statements.AssumeStmt;
 import edu.clemson.cs.rsrg.vcgeneration.proofrules.AbstractProofRuleApplication;
 import edu.clemson.cs.rsrg.vcgeneration.proofrules.ProofRuleApplication;
+import edu.clemson.cs.rsrg.vcgeneration.utilities.Utilities;
 import edu.clemson.cs.rsrg.vcgeneration.vcs.AssertiveCodeBlock;
 import org.stringtemplate.v4.ST;
 import org.stringtemplate.v4.STGroup;
